requirements.txt
requests==2.31.0
flask==3.0.3
apscheduler==3.10.4
python-dotenv==1.0.1
